% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_FeedbackServer.R
\name{FeedbackServer}
\alias{FeedbackServer}
\title{Server for feedback module}
\usage{
FeedbackServer(id = "feedback")
}
\value{
moduleServer for feedback module
}
\description{
Server for feedback module
}
