% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_FeedbackUI.R
\name{FeedbackUI}
\alias{FeedbackUI}
\title{UI for feedback module}
\usage{
FeedbackUI(id = "feedback")
}
\arguments{
\item{strID}{}
}
\value{
tabItem() for feedback UI
}
\description{
UI for feedback module
}
