% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_HomeServer.R
\name{HomeServer}
\alias{HomeServer}
\title{Server for home tab}
\usage{
HomeServer(id = "Home")
}
\arguments{
\item{strID}{Tab ID}
}
\value{
A Shiny module server function for the home tab.
}
\description{
Server for home tab
}
