% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_OptionsServer.R
\name{OptionsServer}
\alias{OptionsServer}
\title{Server for options module}
\usage{
OptionsServer(id = "options")
}
\value{
moduleServer for options module - currently and empty shell
}
\description{
Server for options module
}
