% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_OptionsThemeSwitcherServer.R
\name{OptionsThemeSwitcherServer}
\alias{OptionsThemeSwitcherServer}
\title{Theme Switcher Module Server}
\usage{
OptionsThemeSwitcherServer(id = "themeswitcher")
}
\value{
moduleServer() for theme switcher module
}
\description{
Theme Switcher Module Server
}
