% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_OptionsThemeSwitcherUI.R
\name{OptionsThemeSwithcherUI}
\alias{OptionsThemeSwithcherUI}
\title{Theme switcher module UI}
\usage{
OptionsThemeSwithcherUI(id = "themeswitcher", defaultTheme = "grey_light")
}
\value{
UI for theme selector
}
\description{
Theme switcher module UI
}
