% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SimulationServer.R
\name{SimulationServer}
\alias{SimulationServer}
\title{Server for simulation module}
\usage{
SimulationServer(id = "simulation")
}
\value{
moduleServer for simulation module - currently an empty shell
}
\description{
Server for simulation module
}
